requirements.txt
streamlit==1.14.0
tensorflow-cpu==2.9.1
protobuf==3.19.6
pillow
numpy<2
altair<5
